requirements.txt
streamlit
pandas
python-jobspy
streamlit-extras
